// daisy_scout_web/db/models/leader-2025-data.tsx
export type Leader2025Data = {
  eventId: string;
  matchType: string;
  matchNumber: number;
  source: string | null;
  scoutName: string;
  allianceColor: string;
  alliancePosition: number;
  driverAbility: number | null;
  sourceTime: number | null;
  primaryRole: string | null;
  netShotsMade: number;
  netShotsMissed: number;
  otherNotes: string | null;
  createdAt: Date | null | undefined;
  updatedAt: Date | null | undefined;
};